SocialNetwork/requirements.txt
argcomplete==1.9.3
astroid==2.1.0
awscli==1.16.84
boto3==1.9.74
botocore==1.12.74
certifi==2018.11.29
cfn-flip==1.1.0.post1
chardet==3.0.4
clearbit==0.1.7
Click==7.0
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.8
django-extensions==2.1.4
django-hstore==1.5a0
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
djangorestframework==3.9.0
djangorestframework-hstore==1.3
djangorestframework-jwt==1.11.0
docutils==0.14
durationpy==0.5
email-hunter-python==1.1.0
FullContact.py==0.0.6
future==0.16.0
hjson==3.0.1
idna==2.8
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
openapi-codec==1.3.2
placebo==0.8.2
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pyasn1==0.4.5
PyJWT==1.7.1
pylint==2.2.2
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2018.7
PyYAML==3.13
requests==2.21.0
rsa==3.4.2
s3transfer==0.1.13
simplejson==3.16.0
six==1.12.0
toml==0.10.0
tqdm==4.19.1
troposphere==2.3.4
typed-ast==1.1.1
Unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.10.11
wsgi-request-logger==0.4.6
yapf==0.25.0
zappa==0.47.1
argcomplete==1.9.3
astroid==2.1.0
boto==2.49.0
boto3==1.9.74
botocore==1.12.74
certifi==2018.11.29
cfn-flip==1.1.0.post1
chardet==3.0.4
clearbit==0.1.7
Click==7.0
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.8
django-extensions==2.1.4
django-hstore==1.5a0
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
django-storages==1.7.1
djangorestframework==3.9.0
djangorestframework-hstore==1.3
djangorestframework-jwt==1.11.0
docutils==0.14
durationpy==0.5
email-hunter-python==1.1.0
FullContact.py==0.0.6
future==0.16.0
hjson==3.0.1
idna==2.8
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
openapi-codec==1.3.2
placebo==0.8.2
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pyasn1==0.4.5
PyJWT==1.7.1
pylint==2.2.2
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2018.7
PyYAML==3.13
requests==2.21.0
rsa==3.4.2
s3transfer==0.1.13
simplejson==3.16.0
six==1.12.0
toml==0.10.0
tqdm==4.19.1
troposphere==2.3.4
typed-ast==1.1.1
Unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.10.11
wsgi-request-logger==0.4.6
yapf==0.25.0
zappa==0.47.1
